Use double precision for x in proposeTrajectory
The Euler deterministic-sweep integrator (Task 18) advances x by
dx = -alpha*x*(1-x)*dt per step, which for alpha=200, dt=2.5e-8,
x near 1 gives |dx| ~ 2.5e-12. Float ULP near 1.0 is ~6e-8, so
the increment was silently rounded away and x stayed pinned at
1.0 forever, hitting the 5e8 step safety limit.

Changed 'float x' to 'double x' in proposeTrajectory and cast to
float at buffer-write time. The buffer is still float[1024] for
storage compactness; only the in-flight integrator state needs
double precision. Closed-form mode is byte-identical to the
prior behavior (verified via diff on the smoke test).

+++ src/core/discoalFunctions.c
@@ -1772,8 +1772,8 @@ double initialFreq, double *finalFreq, double alpha, double f0, double currentTi
 	double Nmax, localNextTime,localCurrentTime, currentSizeRatio;
 	int i, insweepphase;
 	long int j;
-	float x;
-	
+	double x;
+
 	// For sweep simulations, write directly to a temporary file
 	char tempFilename[256];
 	snprintf(tempFilename, sizeof(tempFilename), "/tmp/discoal_traj_%d_%ld_%d.tmp", 
@@ -1850,7 +1850,7 @@ double initialFreq, double *finalFreq, double alpha, double f0, double currentTi
 			}
 			
 			// Write to buffer
-			writeBuffer[bufferPos++] = x;
+			writeBuffer[bufferPos++] = (float)x;
 			if (bufferPos >= 1024) {
 				// Flush buffer to file
 				fwrite(writeBuffer, sizeof(float), bufferPos, trajFile);
